Add fn Editor::cmd_action
