+++ crates/kas-widgets/src/edit/edit_field.rs
@@ -9,15 +9,12 @@ use super::*;
 use kas::event::Ime;
 use kas::event::components::TextInputAction;
 use kas::event::{CursorIcon, ElementState, FocusSource, PhysicalKey};
-use kas::geom::Vec2;
 use kas::messages::{ReplaceSelectedText, SetValueText};
 use kas::prelude::*;
 use kas::text::{Effect, EffectFlags, NotReady};
 use kas::theme::TextClass;
 use std::fmt::{Debug, Display};
-use std::ops::Range;
 use std::str::FromStr;
-use unicode_segmentation::{GraphemeCursor, UnicodeSegmentation};
 
 #[impl_self]
 mod EditField {
@@ -671,29 +668,6 @@ impl<G: EditGuard> EditField<G> {
         self
     }
 
-    /// Request key focus, if we don't have it or IME
-    fn request_key_focus(&self, cx: &mut EventState, source: FocusSource) {
-        if !self.has_key_focus && !self.current.is_ime_enabled() {
-            cx.request_key_focus(self.id(), source);
-        }
-    }
-
-    fn trim_paste(&self, text: &str) -> Range<usize> {
-        let mut end = text.len();
-        if !self.multi_line() {
-            // We cut the content short on control characters and
-            // ignore them (preventing line-breaks and ignoring any
-            // actions such as recursive-paste).
-            for (i, c) in text.char_indices() {
-                if c < '\u{20}' || ('\u{7f}'..='\u{9f}').contains(&c) {
-                    end = i;
-                    break;
-                }
-            }
-        }
-        0..end
-    }
-
     fn control_key(
         &mut self,
         cx: &mut EventCx,
@@ -701,296 +675,21 @@ impl<G: EditGuard> EditField<G> {
         cmd: Command,
         code: Option<PhysicalKey>,
     ) -> Result<IsUsed, NotReady> {
-        let editable = self.editable;
-        let mut shift = cx.modifiers().shift_key();
-        let mut buf = [0u8; 4];
-        let cursor = self.selection.edit_index();
-        let len = self.text.str_len();
-        let multi_line = self.multi_line();
-        let selection = self.selection.range();
-        let have_sel = selection.end > selection.start;
-        let string;
-
-        enum Action<'a> {
-            None,
-            Deselect,
-            Activate,
-            Insert(&'a str, EditOp),
-            Delete(Range<usize>, EditOp),
-            Move(usize, Option<f32>),
-            UndoRedo(bool),
+        let action = self.editor.cmd_action(cx, cmd)?;
+        if matches!(action, CmdAction::Unused) {
+            return Ok(Unused);
         }
 
-        let action = match cmd {
-            Command::Escape | Command::Deselect if !selection.is_empty() => Action::Deselect,
-            Command::Activate => Action::Activate,
-            Command::Enter if shift || !multi_line => Action::Activate,
-            Command::Enter if editable && multi_line => {
-                Action::Insert('\n'.encode_utf8(&mut buf), EditOp::KeyInput)
-            }
-            // NOTE: we might choose to optionally handle Tab in the future,
-            // but without some workaround it prevents keyboard navigation.
-            // Command::Tab => Action::Insert('\t'.encode_utf8(&mut buf), EditOp::Insert),
-            Command::Left | Command::Home if !shift && have_sel => {
-                Action::Move(selection.start, None)
-            }
-            Command::Left if cursor > 0 => {
-                let mut cursor = GraphemeCursor::new(cursor, len, true);
-                cursor
-                    .prev_boundary(self.text.text(), 0)
-                    .unwrap()
-                    .map(|index| Action::Move(index, None))
-                    .unwrap_or(Action::None)
-            }
-            Command::Right | Command::End if !shift && have_sel => {
-                Action::Move(selection.end, None)
-            }
-            Command::Right if cursor < len => {
-                let mut cursor = GraphemeCursor::new(cursor, len, true);
-                cursor
-                    .next_boundary(self.text.text(), 0)
-                    .unwrap()
-                    .map(|index| Action::Move(index, None))
-                    .unwrap_or(Action::None)
-            }
-            Command::WordLeft if cursor > 0 => {
-                let mut iter = self.text.text()[0..cursor].split_word_bound_indices();
-                let mut p = iter.next_back().map(|(index, _)| index).unwrap_or(0);
-                while self.text.text()[p..]
-                    .chars()
-                    .next()
-                    .map(|c| c.is_whitespace())
-                    .unwrap_or(false)
-                {
-                    if let Some((index, _)) = iter.next_back() {
-                        p = index;
-                    } else {
-                        break;
-                    }
-                }
-                Action::Move(p, None)
-            }
-            Command::WordRight if cursor < len => {
-                let mut iter = self.text.text()[cursor..]
-                    .split_word_bound_indices()
-                    .skip(1);
-                let mut p = iter.next().map(|(index, _)| cursor + index).unwrap_or(len);
-                while self.text.text()[p..]
-                    .chars()
-                    .next()
-                    .map(|c| c.is_whitespace())
-                    .unwrap_or(false)
-                {
-                    if let Some((index, _)) = iter.next() {
-                        p = cursor + index;
-                    } else {
-                        break;
-                    }
-                }
-                Action::Move(p, None)
-            }
-            // Avoid use of unused navigation keys (e.g. by ScrollComponent):
-            Command::Left | Command::Right | Command::WordLeft | Command::WordRight => Action::None,
-            Command::Up | Command::Down if multi_line => {
-                let x = match self.edit_x_coord {
-                    Some(x) => x,
-                    None => self
-                        .text
-                        .text_glyph_pos(cursor)?
-                        .next_back()
-                        .map(|r| r.pos.0)
-                        .unwrap_or(0.0),
-                };
-                let mut line = self.text.find_line(cursor)?.map(|r| r.0).unwrap_or(0);
-                // We can tolerate invalid line numbers here!
-                line = match cmd {
-                    Command::Up => line.wrapping_sub(1),
-                    Command::Down => line.wrapping_add(1),
-                    _ => unreachable!(),
-                };
-                const HALF: usize = usize::MAX / 2;
-                let nearest_end = match line {
-                    0..=HALF => len,
-                    _ => 0,
-                };
-                self.text
-                    .line_index_nearest(line, x)?
-                    .map(|index| Action::Move(index, Some(x)))
-                    .unwrap_or(Action::Move(nearest_end, None))
-            }
-            Command::Home if cursor > 0 => {
-                let index = self.text.find_line(cursor)?.map(|r| r.1.start).unwrap_or(0);
-                Action::Move(index, None)
-            }
-            Command::End if cursor < len => {
-                let index = self.text.find_line(cursor)?.map(|r| r.1.end).unwrap_or(len);
-                Action::Move(index, None)
-            }
-            Command::DocHome if cursor > 0 => Action::Move(0, None),
-            Command::DocEnd if cursor < len => Action::Move(len, None),
-            // Avoid use of unused navigation keys (e.g. by ScrollComponent):
-            Command::Home | Command::End | Command::DocHome | Command::DocEnd => Action::None,
-            Command::PageUp | Command::PageDown if multi_line => {
-                let mut v = self
-                    .text
-                    .text_glyph_pos(cursor)?
-                    .next_back()
-                    .map(|r| r.pos.into())
-                    .unwrap_or(Vec2::ZERO);
-                if let Some(x) = self.edit_x_coord {
-                    v.0 = x;
-                }
-                const FACTOR: f32 = 2.0 / 3.0;
-                let mut h_dist = f32::conv(self.text.rect().size.1) * FACTOR;
-                if cmd == Command::PageUp {
-                    h_dist *= -1.0;
-                }
-                v.1 += h_dist;
-                Action::Move(self.text.text_index_nearest(v)?, Some(v.0))
-            }
-            Command::Delete | Command::DelBack if editable && have_sel => {
-                Action::Delete(selection.clone(), EditOp::Delete)
-            }
-            Command::Delete if editable => GraphemeCursor::new(cursor, len, true)
-                .next_boundary(self.text.text(), 0)
-                .unwrap()
-                .map(|next| Action::Delete(cursor..next, EditOp::Delete))
-                .unwrap_or(Action::None),
-            Command::DelBack if editable => {
-                // We always delete one code-point, not one grapheme cluster:
-                let prev = self.text.text()[0..cursor]
-                    .char_indices()
-                    .next_back()
-                    .map(|(i, _)| i)
-                    .unwrap_or(0);
-                Action::Delete(prev..cursor, EditOp::Delete)
-            }
-            Command::DelWord if editable => {
-                let next = self.text.text()[cursor..]
-                    .split_word_bound_indices()
-                    .nth(1)
-                    .map(|(index, _)| cursor + index)
-                    .unwrap_or(len);
-                Action::Delete(cursor..next, EditOp::Delete)
-            }
-            Command::DelWordBack if editable => {
-                let prev = self.text.text()[0..cursor]
-                    .split_word_bound_indices()
-                    .next_back()
-                    .map(|(index, _)| index)
-                    .unwrap_or(0);
-                Action::Delete(prev..cursor, EditOp::Delete)
-            }
-            Command::SelectAll => {
-                self.selection.set_sel_index(0);
-                shift = true; // hack
-                Action::Move(len, None)
-            }
-            Command::Cut if editable && have_sel => {
-                cx.set_clipboard((self.text.text()[selection.clone()]).into());
-                Action::Delete(selection.clone(), EditOp::Clipboard)
-            }
-            Command::Copy if have_sel => {
-                cx.set_clipboard((self.text.text()[selection.clone()]).into());
-                Action::None
-            }
-            Command::Paste if editable => {
-                if let Some(content) = cx.get_clipboard() {
-                    let range = self.trim_paste(&content);
-                    string = content;
-                    Action::Insert(&string[range], EditOp::Clipboard)
-                } else {
-                    Action::None
-                }
-            }
-            Command::Undo | Command::Redo if editable => Action::UndoRedo(cmd == Command::Redo),
-            _ => return Ok(Unused),
-        };
+        self.prepare_and_scroll(cx, true);
 
-        // We can receive some commands without key focus as a result of
-        // selection focus. Request focus on edit actions (like Command::Cut).
-        if !matches!(action, Action::None | Action::Deselect) {
-            self.request_key_focus(cx, FocusSource::Synthetic);
-        }
-
-        if !matches!(action, Action::None) {
-            self.cancel_selection_and_ime(cx);
-        }
-
-        let edit_action = match action {
-            Action::None => return Ok(Used),
-            Action::Deselect | Action::Move(_, _) => Some(EditOp::Cursor),
-            Action::Activate | Action::UndoRedo(_) => None,
-            Action::Insert(_, edit) | Action::Delete(_, edit) => Some(edit),
-        };
-        self.save_undo_state(edit_action);
-
-        let mut force_set_offset = false;
-        let result = match action {
-            Action::None => EditAction::None,
-            Action::Deselect => {
-                self.selection.set_empty();
-                cx.redraw();
-                EditAction::None
-            }
-            Action::Activate => {
-                force_set_offset = true;
-                EditAction::Activate
-            }
-            Action::Insert(s, _) => {
-                let mut index = cursor;
-                let range = if have_sel {
-                    index = selection.start;
-                    selection.clone()
-                } else {
-                    index..index
-                };
-                self.text.replace_range(range, s);
-                self.selection.set_cursor(index + s.len());
-                self.edit_x_coord = None;
-                EditAction::Edit
-            }
-            Action::Delete(sel, _) => {
-                self.text.replace_range(sel.clone(), "");
-                self.selection.set_cursor(sel.start);
-                self.edit_x_coord = None;
-                EditAction::Edit
-            }
-            Action::Move(index, x_coord) => {
-                self.selection.set_edit_index(index);
-                if !shift {
-                    self.selection.set_empty();
-                } else {
-                    self.set_primary(cx);
-                }
-                self.edit_x_coord = x_coord;
-                force_set_offset = true;
-                cx.redraw();
-                EditAction::None
-            }
-            Action::UndoRedo(redo) => {
-                if let Some((text, cursor)) = self.editor.undo_stack.undo_or_redo(redo) {
-                    if self.editor.text.set_str(text) {
-                        self.editor.edit_x_coord = None;
-                        self.editor.error_state = false;
-                    }
-                    self.selection = (*cursor).into();
-                    EditAction::Edit
-                } else {
-                    return Ok(Used);
-                }
-            }
-        };
-
-        self.prepare_and_scroll(cx, force_set_offset);
-
-        Ok(match result {
-            EditAction::None => Used,
-            EditAction::Activate => {
+        Ok(match action {
+            CmdAction::Unused => Unused,
+            CmdAction::Used | CmdAction::Cursor => Used,
+            CmdAction::Activate => {
                 cx.depress_with_key(&self, code);
                 G::activate(self, cx, data)
             }
-            EditAction::Edit => {
+            CmdAction::Edit => {
                 G::edit(self, cx, data);
                 Used
             }
